Refactor: Input, Button 컴포넌트 root단으로 이동
모든 컴포넌트에서 참조할 수 있도록 위치 변경
