+++ src/app/user/signup/component/InputID.tsx
@@ -3,7 +3,7 @@ import { SignUpRequestState } from "@/app/recoil/user";
 import { SignUpRequestType } from "@/app/types/user";
 import { ChangeEvent } from "react";
 import { useRecoilState } from "recoil";
-import Input from "../components/common/Input";
+import Input from "../../../component/common/Input";
 
 export default function InputID() {
   const [signUpRequest, setSignUpRequest] =
